% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{aplicar_paleta_victorgm}
\alias{aplicar_paleta_victorgm}
\title{Fijar estilo de gráfico estándar para VictorGM, con algunos parámetros opcionales}
\usage{
aplicar_paleta_victorgm(.x, .n_colors = NULL)
}
\arguments{
\item{.x}{Un objeto ggplot al que se quiere aplicar la paleta.}

\item{.n_colors}{Número entero opcional. Si se especifica, interpola la paleta para generar el número de colores indicado. Por defecto es \code{NULL}, usando la paleta estándar.}
}
\value{
El objeto ggplot modificado con la escala de colores aplicada.
}
\description{
Aplica la paleta de colores de victorgm a un gráfico ggplot.
}
