% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{colores_alternativos}
\alias{colores_alternativos}
\title{Paleta de colores alternativos de victorgm}
\usage{
colores_alternativos()
}
\value{
Un vector de caracteres con códigos de colores hexadecimales.
}
\description{
Devuelve un vector de caracteres con los códigos hexadecimales de una paleta de colores alternativa.
}
