% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{colores_victorgm}
\alias{colores_victorgm}
\title{Paleta de colores principal de victorgm}
\usage{
colores_victorgm()
}
\value{
Un vector de caracteres con códigos de colores hexadecimales.
}
\description{
Devuelve un vector de caracteres con los códigos hexadecimales de la paleta de colores principal.
}
