% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{colores_victorgm_paired}
\alias{colores_victorgm_paired}
\title{Paleta de colores "paired" de victorgm}
\usage{
colores_victorgm_paired()
}
\value{
Un vector de caracteres con códigos de colores hexadecimales.
}
\description{
Devuelve un vector de caracteres con los códigos hexadecimales de la paleta de colores "paired" (pares de colores relacionados).
}
