% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{colores_web}
\alias{colores_web}
\title{Paleta de colores para web de victorgm}
\usage{
colores_web()
}
\value{
Un vector de caracteres con 16 códigos de colores hexadecimales.
}
\description{
Devuelve un vector de 16 colores optimizados para publicaciones en el blog.
Organizados en 5 niveles tonales de púrpura/dorado/verde más un neutro oscuro.
}
