% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{crear_agrupacion}
\alias{crear_agrupacion}
\title{Crear agrupación de columnas para tabla gt}
\usage{
crear_agrupacion(label, columns)
}
\arguments{
\item{label}{Etiqueta de la agrupación}

\item{columns}{Vector de nombres de columnas a agrupar}
}
\value{
Lista con label y columns
}
\description{
Crear agrupación de columnas para tabla gt
}
