% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_acumulado}
\alias{get_acumulado}
\title{Calcula el valor acumulado a los periodos especificados. Devuelve el dataframe original con una columna adicional llamada \verb{acumulado_(.periodos_atras)_(frecuencia)} que contiene el cálculo de la suma de los últimos periodos.}
\usage{
get_acumulado(
  .datos_df,
  .periodos_atras = NULL,
  .frecuencia = "mensual",
  .col_valores = "valores",
  .col_nombres = "nombres",
  .col_flujo = NULL,
  .col_fecha = "fecha"
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos sobre los que se quiere calcular el acumulado.}

\item{.periodos_atras}{Número. Número de periodos sobre el que se quiera calcular el acumulado. Por defecto, 1 año (en el caso de datos semanales se aproximará con 52 semanas, y en el caso de datos diarios con 365 días sin tener en cuenta años bisiestos).}

\item{.frecuencia}{Cadena de caracteres. Indica la frecuencia de los datos. Permite \code{"anual"}, \code{"semestral"}, \code{"trimestral"}, \code{"mensual"}, \code{"semanal"} y \code{"diaria"}. Por defecto, \code{"mensual"}.}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores sobre los que se quiere calcular el acumulado. Por defecto, \code{"valores"}.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere calcular el acumulado. Por ejemplo, si se quiere calcular el acumulado para cada provincia o para cada flujo se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("flujo", "pais")}). Por defecto, \code{"nombres"}. En su ausencia señalar \code{.col_nombres = NULL}.}

\item{.col_flujo}{Cadena de caracteres. Establece el nombre de la columna que determina el sentido del flujo. Si no existiese esta columna y todas las observaciones reflejasen el mismo tipo de flujo permite establecer \code{.col_flujo = NULL}, que es su valor por defecto. Nótese que también se puede establecer la columna flujo en \code{.col_nombres} con el mismo efecto.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Se entiende que habrá un dato mensual. Por defecto, \code{"fecha"}.}
}
\description{
Calcula el valor acumulado a los periodos especificados. Devuelve el dataframe original con una columna adicional llamada \verb{acumulado_(.periodos_atras)_(frecuencia)} que contiene el cálculo de la suma de los últimos periodos.
}
