% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_comercio.R
\name{get_availability}
\alias{get_availability}
\title{Obtener disponibilidad de datos TRAINS}
\usage{
get_availability()
}
\value{
Tibble con informaci칩n de disponibilidad: iso3_code, country_code, country_name, year, nomen_code, n_preferential_agreement, partner_list, specific_duty, last_update.
}
\description{
Obtiene informaci칩n sobre la disponibilidad de datos en el dataset TRAINS de UNCTAD.
No requiere par치metros. Funci칩n adaptada del paquete witstrainsr.
}
\examples{
\dontrun{
availability <- get_availability()
}
}
