% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_cobertura}
\alias{get_cobertura}
\title{Calcula la cobertura. Devuelve el dataframe original, pero en la columna \code{.col_flujo} añade una observación 'cobertura' que divide \code{E} entre \code{I}.}
\usage{
get_cobertura(
  .datos_df,
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_nombres = "nombres",
  .col_valores = "valores",
  .export = "E",
  .import = "I",
  .cobertura = "Cobertura",
  .en_tanto_por_uno = FALSE
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación sobre los que se quiere calcular la cobertura.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el tipo de flujo. El valor en esta columna deberá ser \code{E} para las exportaciones e \code{I} para las importaciones.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere calcular la cobertura. Por ejemplo, si se quiere calcular la cobertura para cada provincia se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("flujo", "pais")}). Por defecto, \code{"nombres"}. En su ausencia señalar \code{.col_nombres = NULL}.}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores. Por defecto, \code{"valores"}.}

\item{.export}{Nombre que denota las exportaciones en la columna .col_flujo. Por defecto, \code{"E"}.}

\item{.import}{Nombre que denota las importaciones en la columna .col_flujo. Por defecto, \code{"I"}.}

\item{.cobertura}{Nombre que denota la cobertura resultante en la columna .col_flujo. Por defecto, \code{"Cobertura"}.}

\item{.en_tanto_por_uno}{Valor lógico. Si se establece en \code{TRUE} devuelve los datos en tanto por uno y en caso contrario en tanto por cien. Por defecto, en tanto por cien.}
}
\description{
Calcula la cobertura. Devuelve el dataframe original, pero en la columna \code{.col_flujo} añade una observación 'cobertura' que divide \code{E} entre \code{I}.
}
