% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_contribucion}
\alias{get_contribucion}
\title{Calcula la contribución al crecimiento en los periodos especificados. Devuelve el dataframe original con una columna adicional llamada \verb{contribucion_(.periodos_atras)_(periodos)} que contiene el cálculo teniendo en cuenta la tasa de variación y el peso.}
\usage{
get_contribucion(
  .datos_df,
  .periodos_atras = NULL,
  .frecuencia = "mensual",
  .col_valores = "valores",
  .col_nombres = "nombres",
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_filtro = NULL,
  .filtro = NULL,
  .en_tanto_por_uno = FALSE
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos sobre los que se quiere calcular la contribución.}

\item{.periodos_atras}{Número. Número de periodos sobre el que se quiera calcular el acumulado. Por defecto, 1 año (en el caso de datos semanales se aproximará con 52 semanas, y en el caso de datos diarios con 365 días sin tener en cuenta años bisiestos).}

\item{.frecuencia}{Cadena de caracteres. Indica la frecuencia de los datos. Permite "anual", "semestral", "trimestral", "mensual", "semanal" y "diaria". Por defecto, "mensual".}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores sobre los que se quiere calcular la contribución. Por defecto, "valores".}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere calcular la contribución. Por ejemplo, si se quiere calcular la contribución de un sector a las exportaciones se debería especificar el nombre de la columna que especifique el sector. Admite varias columnas (p.ej. \code{c("flujo", "pais")}). Por defecto, \code{"nombres"}. En su ausencia señalar \code{.col_nombres = NULL}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Se entiende que habrá un dato mensual. Por defecto, "fecha".}

\item{.col_flujo}{Cadena de caracteres. Establece el nombre de la columna que determina el sentido del flujo. Si no existiese esta columna y todas las observaciones reflejan el mismo tipo de flujo permite establecer \code{.col_flujo = NULL}. Por defecto, \code{.col_flujo = "flujo"}.}

\item{.col_filtro}{Opcional. Cadena de caracteres. Establece la columna (adicional) que incorpore datos sobre los que se quiera establecer el filtro indicado por el parámetro \code{.filtro}.}

\item{.filtro}{Opcional. Establece el valor al que se tiene que igualar la columna \code{.col_filtro} para que la observación no sea descartada de inicio (p.ej. esto sirve para calcular la contribución de un sector a las exportaciones de EEUU).}

\item{.en_tanto_por_uno}{Valor lógico. Si se establece en \code{TRUE} devuelve los datos en tanto por uno y en caso contrario en tanto por cien. Por defecto, en tanto por cien.}
}
\description{
Calcula la contribución al crecimiento en los periodos especificados. Devuelve el dataframe original con una columna adicional llamada \verb{contribucion_(.periodos_atras)_(periodos)} que contiene el cálculo teniendo en cuenta la tasa de variación y el peso.
}
