% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_contribucion_acumulada}
\alias{get_contribucion_acumulada}
\title{Calcula la contribución al crecimiento en los últimos \code{.años_atras_contribucion} del acumulado a \code{.periodos_atras_acumulado}. Devuelve un dataframe con 2 columnas: \code{nombres} y \code{valores}, que hacen referencia a la contribución a \code{años_atras_contribucion} a la variación del \code{.flujo} en el último periodo con datos disponibles.}
\usage{
get_contribucion_acumulada(
  .datos_df,
  .años_atras_contribucion = 3,
  .periodos_atras_acumulado = NULL,
  .frecuencia = "mensual",
  .col_valores_contr_acum = "valores",
  .col_nombres_contr_acum = "nombres",
  .col_fecha_contr_acum = "fecha",
  .col_flujo = "flujo",
  .col_filtro = NULL,
  .filtro = NULL,
  .en_tanto_por_uno = FALSE
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos sobre los que se quiere calcular la contribución. Es necesario que comprenda únicamente exportaciones o únicamente importaciones. Preferiblemente incluirá 3 columnas \code{fecha}, \code{nombres} y \code{valores}, y la columna \code{nombres} incluirá datos de la variable sobre la que se quiere calcular la contribución y al nivel de desagregación deseado (p.ej. si se quiere obtener la contribución por sectores nivel 1, \code{datos_df} incluirá el nombre de los sectores nivel 1 en la columna \code{nombres}).}

\item{.años_atras_contribucion}{Número. Número de años sobre el que se quiera calcular la contribución al crecimiento. Por defecto, 3 años.}

\item{.periodos_atras_acumulado}{Número. Número de periodos sobre el que se quiera calcular el acumulado sobre el que establecer la tasa de variación. Por defecto, 1 año (en el caso de datos semanales se aproximará con 52 semanas, y en el caso de datos diarios con 365 días sin tener en cuenta años bisiestos).}

\item{.frecuencia}{Cadena de caracteres. Indica la frecuencia de los datos. Permite \code{"anual"}, \code{"semestral"}, \code{"trimestral"}, \code{"mensual"}, \verb{"semanal"`` y }"diaria"\verb{. Por defecto, }"mensual"`.}

\item{.col_valores_contr_acum}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores sobre los que se quiere calcular el acumulado. Por defecto, \code{"valores"}.}

\item{.col_nombres_contr_acum}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere calcular el acumulado. Por ejemplo, si se quiere calcular el acumulado para cada provincia o para cada flujo se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("flujo", "pais")}). Por defecto, \code{"nombres"}.}

\item{.col_fecha_contr_acum}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Se entiende que habrá un dato mensual. Por defecto, "fecha".}

\item{.col_flujo}{Cadena de caracteres. Establece el nombre de la columna que determina el sentido del flujo. Si no se existe esta columna y todas las observaciones reflejan el mismo tipo de flujo permite establecer \code{.col_flujo = NULL}. Por defecto, \code{"flujo"}.}

\item{.col_filtro}{Opcional. Cadena de caracteres. Establece la columna (adicional) que incorpore datos sobre los que se quiera establecer el filtro indicado por el parámetro \code{.filtro}.}

\item{.filtro}{Opcional. Establece el valor al que se tiene que igualar la columna \code{.col_filtro} para que la observación no sea descartada de inicio (p.ej. esto sirve para calcular la contribución de un sector a las exportaciones de EEUU).}

\item{.en_tanto_por_uno}{Valor lógico. Si se establece en \code{TRUE} devuelve los datos en tanto por uno y en caso contrario en tanto por cien. Por defecto, en tanto por cien.}
}
\description{
Calcula la contribución al crecimiento en los últimos \code{.años_atras_contribucion} del acumulado a \code{.periodos_atras_acumulado}. Devuelve un dataframe con 2 columnas: \code{nombres} y \code{valores}, que hacen referencia a la contribución a \code{años_atras_contribucion} a la variación del \code{.flujo} en el último periodo con datos disponibles.
}
