% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_tipocambio.R
\name{get_conversion}
\alias{get_conversion}
\title{Función para convertir los valores monetarios de una columna a otra divisa en un dataframe que refleje datos de series temporales.}
\usage{
get_conversion(
  .datos_df,
  .from = "EUR",
  .to = "USD",
  .frecuencia = "mensual",
  .col_valores = "valores",
  .col_fecha = "fecha"
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos sobre los que se quiere calcular la conversión.}

\item{.from}{Cadena de caracteres. Código ISO a 3 letras de la divisa de base (base currency). Por defecto, \code{"EUR"}.}

\item{.to}{Cadena de caracteres. Código ISO a 3 letras de la divisa de cuenta (quote currency). Por defecto, \code{"USD"}.}

\item{.frecuencia}{Cadena de caracteres. Frecuencia de los datos. Puede ser, \code{"diaria"}, \code{"mensual"} o \code{"anual"}. Por defecto, \code{"mensual"}.}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna que contiene los datos que se quieren convertir. Por defecto, \code{"valores"}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna que contiene las fechas de los datos que se quieren convertir. Por defecto, \code{"fecha"}.}
}
\description{
Función para convertir los valores monetarios de una columna a otra divisa en un dataframe que refleje datos de series temporales.
}
