% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_comercio.R
\name{get_countries}
\alias{get_countries}
\title{Obtener lista de países TRAINS}
\usage{
get_countries()
}
\value{
Tibble con columnas: iso3_code, country_code, name, is_reporter, is_partner, is_group.
}
\description{
Obtiene nombres, códigos alfanuméricos, códigos ISO3 y otros metadatos de los países
disponibles en el dataset TRAINS de UNCTAD. No requiere parámetros.
Función adaptada del paquete witstrainsr.
}
\examples{
\dontrun{
countries <- get_countries()
}
}
