% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_precios.R
\name{get_descomposicion_tv}
\alias{get_descomposicion_tv}
\title{Partiendo de un dataframe con valor nominal, precios ponderados y flujos reales, devuelve un dataframe con la tasa de variación en términos nominales y la contribución del precio y del volumen.}
\usage{
get_descomposicion_tv(
  .datos_df,
  .meses_atras = 12,
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_valor_estadistico = "valor_estadistico",
  .col_precios_ponderados = "precios_ponderados",
  .col_flujos_reales = "flujos_reales",
  .col_nombres = NULL,
  .remove_na = TRUE
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación sobre los que se quiere calcular las tasas de variación y las contribuciones. La idea es que incluya \code{.col_fecha}, \code{.col_flujo}, \code{.col_valor_estadistico}, \code{.col_precios_ponderados}, \code{.col_flujos_reales} y opcionalmente, \code{.col_nombres}.}

\item{.meses_atras}{Número. Número de meses sobre los que se quiere calcular la tasa de variación. Por defecto, \code{12}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el tipo de flujo. Por defecto, \code{"flujo"}.}

\item{.col_valor_estadistico}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores estadísticos. Por defecto, \code{"valor_estadistico"}.}

\item{.col_precios_ponderados}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los precios ponderados. Por defecto, \code{"precios_ponderados"}.}

\item{.col_flujos_reales}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los flujos reales. Por defecto, \code{"flujos_reales"}.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere agrupar para calcular los índices. Por ejemplo, si se quiere calcular el índice de precios para cada provincia se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("provincia", "pais")}). Por defecto, \code{NULL}.}
}
\description{
Partiendo de un dataframe con valor nominal, precios ponderados y flujos reales, devuelve un dataframe con la tasa de variación en términos nominales y la contribución del precio y del volumen.
}
