% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_precios.R
\name{get_descomposicion_tv_plt}
\alias{get_descomposicion_tv_plt}
\title{Partiendo de un dataframe como el generado con \code{victorgmtools::get_descomposicion_tv}, permite generar un gráfico para analizar la descomposición de la tasa de variación nominal.}
\usage{
get_descomposicion_tv_plt(
  .datos_df,
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_tv_nominal = "tv_nominal",
  .col_contribucion_precios = "contribucion_precios",
  .col_contribucion_volumen = "contribucion_volumen",
  .col_nombres = NULL,
  .flujo_seleccionado = NULL,
  .remove_na = TRUE,
  .accuracy = 0.01,
  .quitar_valores_extremos = TRUE,
  .desviaciones_extremos = 2,
  .legend_position = "bottom",
  .title = NULL,
  .title_size = 18,
  .subtitle = NULL,
  .subtitle_size = 14,
  .caption = NULL,
  .caption_size = 10,
  .estatico_toggle = FALSE,
  .fuente_letra = "Segoe UI"
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación sobre los que se quiere calcular las tasas de variación y las contribuciones. La idea es que incluya \code{.col_fecha}, \code{.col_flujo}, \code{.col_valor_estadistico}, \code{.col_precios_ponderados}, \code{.col_flujos_reales} y opcionalmente, \code{.col_nombres}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el tipo de flujo. Por defecto, \code{"flujo"}.}

\item{.col_tv_nominal}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las tasas de variación en términos nominales. Por defecto, \code{tv_nominal}.}

\item{.col_contribucion_precios}{Cadena de caracteres. Nombre de la columna del dataframe que contiene la contribución de los precios a la tasa de variación. Por defecto, \code{contribucion_precios}.}

\item{.col_contribucion_volumen}{Cadena de caracteres. Nombre de la columna del dataframe que contiene la contribución del peso a la tasa de variación. Por defecto, \code{contribucion_volumen}.}

\item{.col_nombres}{Opcional (sólo en caso de que .datos_df contenga esta columna). Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere agrupar para calcular las contribuciones a las tasas de variación. Por defecto, \code{NULL}.}

\item{.flujo_seleccionado}{Cadena de caracteres. Permite filtrar por flujo, estableciendo por ejemplo, \code{.flujo_seleccionado = "E"}. Por defecto, \code{NULL}, que devuelve ambos gráficos.}

\item{.remove_na}{Valor lógico. Permite eliminar los primero periodos del dataframe en los que las tasas de variación y las contribuciones son \code{NA} o \code{NaN}. Por defecto, \code{TRUE}.}

\item{.quitar_valores_extremos}{Valor lógico. En caso de \code{TRUE} elimina valores extremos de \code{.col_contribucion_precios} y de \code{.col_contribucion_volumen} para que el gráfico siga siendo informativo. Por defecto, \code{TRUE}.}

\item{.estatico_toggle}{Valor lógico. En caso de \code{TRUE} el gráfico no será interactivo. Por defecto, \code{FALSE}.}

\item{.fuente_letra}{Cadena de caracteres. Fuente tipográfica a utilizar en el gráfico. Por defecto, utiliza la fuente: \code{"Segoe UI"}. Otras opciones comunes: \code{"Arial"}, \code{"Times"}, \code{"Courier"}, \code{"Helvetica"}, etc.}
}
\description{
Partiendo de un dataframe como el generado con \code{victorgmtools::get_descomposicion_tv}, permite generar un gráfico para analizar la descomposición de la tasa de variación nominal.
}
