% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_precios.R
\name{get_flujos_reales}
\alias{get_flujos_reales}
\title{Calcula los flujos en términos reales. Devuelve un dataframe con la evolución de los índices de precios y del valor de las exportaciones en términos reales.}
\usage{
get_flujos_reales(
  .datos_df,
  .año_referencia = NULL,
  .año_base = NULL,
  .meses_atras_acumulado = 12,
  .variable_utilizada_cesta = "valor_estadistico",
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_taric = "taric",
  .col_valor_estadistico = "valor_estadistico",
  .col_peso = "peso",
  .col_nombres = NULL,
  .export = "E",
  .import = "I",
  .remove_na = TRUE,
  .metodo = NULL
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación sobre los que se quiere calcular el índice. La idea es que incluya \code{.col_fecha}, \code{.col_flujo}, \code{.col_taric}, \code{.col_valor_estadistico}, \code{.col_peso} y opcionalmente, \code{.col_nombres}. Si se ha obtenido a través de la función \code{get_precios_ponderados}, lo detecta y se ahorra el cálculo.}

\item{.año_referencia}{Número. Año para tomar como referencia y hacer la cesta de bienes. En su defecto (i.e. si \code{NULL}), se realizarán las ponderaciones con una cesta móvil que evoluciona en base a las observaciones de los últimos \code{.meses_atras_acumulado} meses. Por defecto, \code{NULL}.}

\item{.año_base}{Número. Año para tomar como base (toma la primera observación del año en caso de que los datos no sean anuales (i.e. enero si son mensuales)) y establecer índice 100 a los flujos reales. En su defecto (i.e. si \code{NULL}), no se calcula índice sino que \code{flujos_reales} refleja los \code{kilogramos} de la cesta compuesta ficticia. Por defecto, \code{NULL}.}

\item{.meses_atras_acumulado}{Número. Número de meses sobre los que se calcula el acumulado para obtener las cestas. Por defecto, \code{12}.}

\item{.variable_utilizada_cesta}{Cadena de caracteres. Permite \code{"valor_estadistico"} o \code{"peso"}. Por defecto, \code{"valor_estadistico"}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el tipo de flujo. Por defecto, \code{"flujo"}.}

\item{.col_taric}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el codigo TARIC de cada producto. Por defecto, \code{"taric"}.}

\item{.col_valor_estadistico}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores estadísticos. Por defecto, \code{"valor_estadistico"}.}

\item{.col_peso}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el peso. Es importante que estén definidos en la misma unidad. Por defecto, \code{"peso"}.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere agrupar para calcular los índices. Por ejemplo, si se quiere calcular el índice de precios para cada provincia se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("provincia", "pais")}). Por defecto, \code{NULL}.}

\item{.export}{Nombre que denota las exportaciones en la columna \code{.col_flujo}. Por defecto, \code{"E"}.}

\item{.import}{Nombre que denota las importaciones en la columna \code{.col_flujo}. Por defecto, \code{"I"}.}

\item{.metodo}{Si se establece \code{.metodo = "SQL"} usará SQL para una mayor eficiencia. Deseable cuando \code{.datos_df} es grande. Si se establece en \code{NULL} realizará los cálculos usando únicamente R base y el tidyverse. Por defecto, \code{"SQL"}.}
}
\description{
Calcula los flujos en términos reales. Devuelve un dataframe con la evolución de los índices de precios y del valor de las exportaciones en términos reales.
}
