% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_precios.R
\name{get_flujos_reales_duckplyr}
\alias{get_flujos_reales_duckplyr}
\title{Calcula los flujos en términos reales (versión optimizada).}
\usage{
get_flujos_reales_duckplyr(
  .datos_df,
  .año_referencia = NULL,
  .año_base = NULL,
  .meses_atras_acumulado = 12,
  .variable_utilizada_cesta = "valor_estadistico",
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_taric = "taric",
  .col_valor_estadistico = "valor_estadistico",
  .col_peso = "peso",
  .col_nombres = NULL,
  .export = "E",
  .import = "I",
  .remove_na = TRUE,
  .use_duckplyr = TRUE
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación.}

\item{.año_referencia}{Número. Año para tomar como referencia.}

\item{.año_base}{Número. Año para tomar como base.}

\item{.meses_atras_acumulado}{Número. Número de meses para acumulado.}

\item{.variable_utilizada_cesta}{Cadena de caracteres. Variable para cesta.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna de fechas.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna de flujo.}

\item{.col_taric}{Cadena de caracteres. Nombre de la columna TARIC.}

\item{.col_valor_estadistico}{Cadena de caracteres. Nombre de la columna de valores.}

\item{.col_peso}{Cadena de caracteres. Nombre de la columna de peso.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna de nombres.}

\item{.export}{Nombre que denota exportaciones.}

\item{.import}{Nombre que denota importaciones.}

\item{.remove_na}{Valor lógico. Si TRUE, elimina NAs.}

\item{.use_duckplyr}{Valor lógico. Si TRUE, usa duckplyr.}
}
\description{
Calcula los flujos en términos reales (versión optimizada).
}
