% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_media_movil}
\alias{get_media_movil}
\title{Calcula la media móvil. Devuelve el dataframe original con una columna adicional \verb{mm_(.periodos_atras)_(periodo)} que hace referencia a la media móvil a \code{.periodos_atras} periodos.}
\usage{
get_media_movil(
  .datos_df,
  .periodos_atras = NULL,
  .frecuencia = "mensual",
  .col_valores = "valores",
  .col_nombres = "nombres",
  .col_flujo = NULL,
  .col_fecha = "fecha"
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos sobre los que se quiere calcular la media móvil.}

\item{.periodos_atras}{Número. Número de periodos sobre el que se quiera calcular la tasa de variación. Por defecto, 1 año (en el caso de datos semanales se aproximará con 52 semanas, y en el caso de datos diarios con 365 días sin tener en cuenta años bisiestos).}

\item{.frecuencia}{Cadena de caracteres. Indica la frecuencia de los datos. Permite \code{"anual"}, \code{"semestral"}, \code{"trimestral"}, \code{"mensual"}, \code{"semanal"} y \code{"diaria"}. Por defecto, \code{"mensual"}.}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores. Por defecto, "valores".}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres para realizar la agrupación. Permite introducir el nombre de varias columnas. Por ejemplo, si se quiere calcular la media móvil de las exportaciones para cada provincia sobre el total se podrá especificar aquí \verb{.col_nombres = ("flujo", "provincia")}. Por defecto, \code{"nombres"}.}

\item{.col_flujo}{Cadena de caracteres. Establece el nombre de la columna que determina el sentido del flujo. Si no existiese esta columna y todas las observaciones reflejasen el mismo tipo de flujo permite establecer \code{.col_flujo = NULL}, que es su valor por defecto. Nótese que también se puede establecer la columna flujo en \code{.col_nombres} con el mismo efecto.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}
}
\description{
Calcula la media móvil. Devuelve el dataframe original con una columna adicional \verb{mm_(.periodos_atras)_(periodo)} que hace referencia a la media móvil a \code{.periodos_atras} periodos.
}
