% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_comercio.R
\name{get_nomen_codes}
\alias{get_nomen_codes}
\title{Obtener nomenclaturas disponibles en TRAINS}
\usage{
get_nomen_codes()
}
\value{
Tibble con columnas: nomen_code, description.
}
\description{
Obtiene información sobre las nomenclaturas disponibles en el dataset TRAINS de UNCTAD.
No requiere parámetros. Función adaptada del paquete witstrainsr.
}
\examples{
\dontrun{
nomenclatures <- get_nomen_codes()
}
}
