% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_peso}
\alias{get_peso}
\title{Calcula el peso sobre el total del flujo. Devuelve un dataframe como el original, pero con una columna adiciona: \code{peso}, que hace referencia al peso del flujo sobre el total de \code{col.nombres} en la fecha indicada.}
\usage{
get_peso(
  .datos_df,
  .col_valores = "valores",
  .col_nombres = "nombres",
  .col_flujo = "flujo",
  .col_fecha = "fecha",
  .en_tanto_por_uno = FALSE
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos.}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna con valores. Por defecto, \code{"valores"}.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna con nombres. Por defecto, \code{"nombres"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna de flujo. Si es NULL, se ignora. Por defecto, \code{"flujo"}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna de fecha. Por defecto, \code{"fecha"}.}

\item{.en_tanto_por_uno}{Valor lógico. Si se establece en \code{TRUE} devuelve los datos en tanto por uno y en caso contrario en tanto por cien. Por defecto, en tanto por cien.}
}
\description{
Calcula el peso sobre el total del flujo. Devuelve un dataframe como el original, pero con una columna adiciona: \code{peso}, que hace referencia al peso del flujo sobre el total de \code{col.nombres} en la fecha indicada.
}
