% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_precios.R
\name{get_precios_ponderados}
\alias{get_precios_ponderados}
\title{Calcula índices de precios. Devuelve un dataframe con la evolución de los índices y (en su caso) de la relación real de intercambio (RRI).}
\usage{
get_precios_ponderados(
  .datos_df,
  .año_referencia = NULL,
  .meses_atras_acumulado = 12,
  .variable_utilizada_cesta = "valor_estadistico",
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_taric = "taric",
  .col_valor_estadistico = "valor_estadistico",
  .col_peso = "peso",
  .col_nombres = NULL,
  .flujo_seleccionado = NULL,
  .export = "E",
  .import = "I",
  .get_rri = TRUE,
  .remove_na = TRUE,
  .metodo = NULL
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación sobre los que se quiere calcular el índice. La idea es que incluya \code{.col_fecha}, \code{.col_flujo}, \code{.col_taric}, \code{.col_valor_estadistico}, \code{.col_peso} y opcionalmente, \code{.col_nombres}.}

\item{.año_referencia}{Número. Año para tomar como referencia y hacer la cesta de bienes. En su defecto (i.e. si \code{NULL}), se realizarán las ponderaciones con una cesta móvil que evoluciona en base a las observaciones de los últimos \code{.meses_atras_acumulado} meses. Por defecto, \code{NULL}.}

\item{.meses_atras_acumulado}{Número. Número de meses sobre los que se calcula el acumulado para obtener las cestas. Por defecto, \code{12}.}

\item{.variable_utilizada_cesta}{Cadena de caracteres. Permite \code{"valor_estadistico"} o \code{"peso"}. Por defecto, \code{"valor_estadistico"}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el tipo de flujo. Por defecto, \code{"flujo"}.}

\item{.col_taric}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el codigo TARIC de cada producto. Por defecto, \code{"taric"}.}

\item{.col_valor_estadistico}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores estadísticos. Por defecto, \code{"valor_estadistico"}.}

\item{.col_peso}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el peso. Es importante que estén definidos en la misma unidad. Por defecto, \code{"peso"}.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere agrupar para calcular los índices. Por ejemplo, si se quiere calcular el índice de precios para cada provincia se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("provincia", "pais")}). Por defecto, \code{NULL}.}

\item{.export}{Nombre que denota las exportaciones en la columna \code{.col_flujo}. Por defecto, \code{"E"}.}

\item{.import}{Nombre que denota las importaciones en la columna \code{.col_flujo}. Por defecto, \code{"I"}.}

\item{.get_rri}{Valor lógico. En caso de \code{get_rri = TRUE}, devuelve una columna adicional con la RRI (i.e.cociente del índice de precios de exportación y de importación). Por defecto, \code{TRUE}.}

\item{.metodo}{Si se establece \code{.metodo = "SQL"} usará SQL para una mayor eficiencia. Deseable cuando \code{.datos_df} es grande. Si se establece en \code{NULL} realizará los cálculos usando únicamente R base y el tidyverse. Por defecto, \code{NULL}.}
}
\description{
Calcula índices de precios. Devuelve un dataframe con la evolución de los índices y (en su caso) de la relación real de intercambio (RRI).
}
