% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_precios.R
\name{get_rri_plt}
\alias{get_rri_plt}
\title{Permite generar un gráfico para analizar la evolución del precio de un producto.}
\usage{
get_rri_plt(
  .datos_df,
  .año_referencia = NULL,
  .meses_atras_acumulado = 12,
  .variable_utilizada_cesta = "valor_estadistico",
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_taric = "taric",
  .col_valor_estadistico = "valor_estadistico",
  .col_peso = "peso",
  .col_nombres = NULL,
  .nombres_por_facets = FALSE,
  .export = "E",
  .import = "I",
  .flujo_seleccionado = NULL,
  .accuracy = 0.01,
  .legend_position = "bottom",
  .title = NULL,
  .title_size = 18,
  .subtitle = NULL,
  .subtitle_size = 14,
  .caption = NULL,
  .caption_size = 10,
  .estatico_toggle = FALSE,
  .representar_indices = TRUE,
  .fuente_letra = "Segoe UI"
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación. La idea es que incluya \code{.col_fecha}, \code{.col_flujo}, \code{.col_taric}, \code{.col_valor_estadistico}, \code{.col_peso} y opcionalmente, \code{.col_nombres}. En caso de que el dataframe provenga de la función \code{get_precios_ponderados} se ahorrará los cálculos intermedios.}

\item{.año_referencia}{Número. Año para tomar como referencia y hacer la cesta de bienes. En su defecto (i.e. si \code{NULL}), se realizarán las ponderaciones con una cesta móvil que evoluciona en base a las observaciones de los últimos \code{.meses_atras_acumulado} meses. Por defecto, \code{NULL}.}

\item{.meses_atras_acumulado}{Número. Número de meses sobre los que se calcula el acumulado para obtener las cestas. Por defecto, \code{12}.}

\item{.variable_utilizada_cesta}{Cadena de caracteres. Permite \code{"valor_estadistico"} o \code{"peso"}. Por defecto, \code{"valor_estadistico"}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el tipo de flujo. Por defecto, \code{"flujo"}.}

\item{.col_taric}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el codigo TARIC de cada producto. Por defecto, \code{"taric"}.}

\item{.col_valor_estadistico}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores estadísticos. Por defecto, \code{"valor_estadistico"}.}

\item{.col_peso}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el peso. Es importante que estén definidos en la misma unidad. Por defecto, \code{"peso"}.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres para hacer distintos gráficos por categorías. Por ejemplo, si se quiere representar la evolución de precios para cada provincia se podrá especificar aquí. Por defecto, \code{NULL}.}

\item{.nombres_por_facets}{Valor lógico. Permite establecer si se quiere agregar por nombres antes de calcular el índice de precios (en caso de \code{FALSE}) o si se quiere representar un gráfico para cada valor de la columna \code{.col_nombres} (en caso de \code{TRUE}). Por defecto, \code{FALSE}.}

\item{.export}{Nombre que denota las exportaciones en la columna \code{.col_flujo}. Por defecto, \code{"E"}.}

\item{.import}{Nombre que denota las importaciones en la columna \code{.col_flujo}. Por defecto, \code{"I"}.}

\item{.flujo_seleccionado}{Cadena de caracteres. Permite filtrar por flujo, estableciendo por ejemplo, \code{.flujo_seleccionado = "E"}. Por defecto, \code{NULL}, que devuelve exportaciones e importaciones.}

\item{.estatico_toggle}{Valor lógico. En caso de \code{TRUE} el gráfico no será interactivo. Por defecto, \code{FALSE}.}

\item{.representar_indices}{Valor lógico. En caso de \code{TRUE} el gráfico representa índices en base 100 para el primer período del gráfico, representando 3 índices: RRI, precio de exportaciones y precio de importaciones. En caso de \code{FALSE} representa únicamente la evolución de la RRI (sin calcular índices). Por defecto, \code{TRUE}.}

\item{.fuente_letra}{Cadena de caracteres. Fuente tipográfica a utilizar en el gráfico. Por defecto, utiliza la fuente: \code{"Segoe UI"}. Otras opciones comunes: \code{"Arial"}, \code{"Times"}, \code{"Courier"}, \code{"Helvetica"}, etc.}
}
\description{
Permite generar un gráfico para analizar la evolución del precio de un producto.
}
