% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_saldo}
\alias{get_saldo}
\title{Calcula el saldo. Devuelve el dataframe original, pero en la columna \code{.col_flujo} añade una observación 'saldo' que resta \code{E} menos \code{I}.}
\usage{
get_saldo(
  .datos_df,
  .col_fecha = "fecha",
  .col_flujo = "flujo",
  .col_nombres = "nombres",
  .col_valores = "valores",
  .export = "E",
  .import = "I",
  .saldo = "Saldo"
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos de importación y exportación sobre los que se quiere calcular el saldo.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Nombre de la columna del dataframe que contiene el tipo de flujo. El valor en esta columna deberá ser \code{E} para las exportaciones e \code{I} para las importaciones.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere calcular el saldo. Por ejemplo, si se quiere calcular el saldo para cada provincia se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("flujo", "pais")}). Por defecto, \code{"nombres"}. En su ausencia señalar \code{.col_nombres = NULL}.}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores. Por defecto, \code{"valores"}.}

\item{.export}{Nombre que denota las exportaciones en la columna .col_flujo. Por defecto, \code{"E"}.}

\item{.import}{Nombre que denota las importaciones en la columna .col_flujo. Por defecto, \code{"I"}.}

\item{.saldo}{Nombre que denota el saldo resultante en la columna .col_flujo. Por defecto, \code{"Saldo"}.}
}
\description{
Calcula el saldo. Devuelve el dataframe original, pero en la columna \code{.col_flujo} añade una observación 'saldo' que resta \code{E} menos \code{I}.
}
