% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_comercio.R
\name{get_tariffs}
\alias{get_tariffs}
\title{Obtener datos de aranceles de WITS/TRAINS}
\usage{
get_tariffs(
  reporter,
  partner = "000",
  product = "all",
  year = "all",
  datatype = "reported",
  verbose = TRUE
)
}
\arguments{
\item{reporter}{Vector de caracteres con códigos ISO de 3 dígitos de países declarantes, o "all".}

\item{partner}{Vector de caracteres con códigos ISO de 3 dígitos de países socios, "all" o "000" (mundo). Por defecto, "000".}

\item{product}{Vector de caracteres con códigos de producto HS6 de 6 dígitos, o "all". Por defecto, "all".}

\item{year}{Vector de caracteres, numérico o entero con años en formato de 4 dígitos, o "all". Por defecto, "all".}

\item{datatype}{Tipo de datos: "reported" (solo aranceles) o "aveestimated" (incluye equivalentes ad valorem). Por defecto, "reported".}

\item{verbose}{Lógico. Si TRUE, muestra mensajes de progreso. Por defecto, TRUE.}
}
\value{
Tibble con columnas: reporter_country_code, partner_country_code, product_code, year, tariff_simple_average, tariff_type, n_total_lines, etc.
}
\description{
Obtiene datos de aranceles desde la base de datos World International Trade Solution (WITS)
del Banco Mundial. Función adaptada del paquete witstrainsr. Los datos están armonizados
a nivel de 6 dígitos del Sistema Armonizado (HS6) e incluyen tasas preferenciales y NMF.
}
\examples{
\dontrun{
get_tariffs(reporter = "ESP", partner = "000", year = 2023)
get_tariffs(reporter = "USA", partner = "CHN", product = "all", year = 2020:2023)
}
}
