% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_calculos.R
\name{get_tasa_variacion_fecha_ref}
\alias{get_tasa_variacion_fecha_ref}
\title{Calcula la tasa de variación de una variable con respecto al valor alcanzado en la fecha seleccionada. Devuelve el dataframe original con una columna adicional llamada \code{tv_respecto_(.fecha_referencia)}.}
\usage{
get_tasa_variacion_fecha_ref(
  .datos_df,
  .fecha_referencia,
  .frecuencia = "mensual",
  .col_valores = "valores",
  .col_nombres = "nombres",
  .col_fecha = "fecha",
  .col_flujo = NULL,
  .en_tanto_por_uno = FALSE,
  .completar_df = FALSE
)
}
\arguments{
\item{.datos_df}{Data frame que contiene los datos sobre los que se quiere calcular la tasa de variación.}

\item{.fecha_referencia}{Fecha. Periodo con respecto al que se quiere obtener la comparación.}

\item{.frecuencia}{Cadena de caracteres. Indica la frecuencia de los datos. Permite \code{"anual"}, \code{"semestral"}, \code{"trimestral"}, \code{"mensual"}, \code{"semanal"} y \code{"diaria"}. Por defecto, \code{"mensual"}.}

\item{.col_valores}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los valores sobre los que se quiere calcular la tasa de variación. Por defecto, \code{"valores"}.}

\item{.col_nombres}{Cadena de caracteres. Nombre de la columna del dataframe que contiene los nombres sobre los que se quiere calcular la tasa de variación. Por ejemplo, si se quiere calcular la tasa de variación para cada provincia o para cada flujo se podrá especificar aquí. Admite varias columnas (p.ej. \code{c("flujo", "pais")}). Por defecto, "nombres". En su ausencia señalar \code{.col_nombres = NULL}.}

\item{.col_fecha}{Cadena de caracteres. Nombre de la columna del dataframe que contiene las fechas (en formato fecha). La fórmula entiende que los datos son mensuales. Por defecto, \code{"fecha"}.}

\item{.col_flujo}{Cadena de caracteres. Establece el nombre de la columna que determina el sentido del flujo. Si no existiese esta columna y todas las observaciones reflejasen el mismo tipo de flujo permite establecer \code{.col_flujo = NULL}, que es su valor por defecto. Nótese que también se puede establecer la columna flujo en \code{.col_nombres} con el mismo efecto.}

\item{.en_tanto_por_uno}{Valor lógico. Si se establece en \code{TRUE} devuelve los datos en tanto por uno y en caso contrario en tanto por cien. Por defecto, en tanto por cien.}

\item{.completar_df}{Valor lógico. Indica si quieres rellenar los datos faltantes con ceros. Por defecto, \code{FALSE}.}
}
\description{
Calcula la tasa de variación de una variable con respecto al valor alcanzado en la fecha seleccionada. Devuelve el dataframe original con una columna adicional llamada \code{tv_respecto_(.fecha_referencia)}.
}
