% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_tipocambio.R
\name{get_tipo_cambio}
\alias{get_tipo_cambio}
\title{Función para obtener el tipo de cambio entre 2 divisas en una fecha específica.}
\usage{
get_tipo_cambio(.from = "EUR", .to = "USD", .fecha = Sys.Date())
}
\arguments{
\item{.from}{Cadena de caracteres. Código ISO a 3 letras de la divisa de base (base currency). Por defecto, \code{"EUR"}.}

\item{.to}{Cadena de caracteres. Código ISO a 3 letras de la divisa de cuenta (quote currency). Por defecto, \code{"USD"}.}

\item{.fecha}{Fecha en formato "\%Y-\%m-\%d" para la que se quiere obtener el tipo de cambio. Por defecto, hoy.}
}
\description{
Función para obtener el tipo de cambio entre 2 divisas en una fecha específica.
}
