% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_tipocambio.R
\name{get_tipo_cambio_df}
\alias{get_tipo_cambio_df}
\title{Funci贸n para obtener un dataframe con datos del tipo de cambio entre 2 divisas en un periodo especificado. Devuelve el dataframe inicial con una columna adicional \verb{tipo_cambio_".from_currency"_".to_currency"}, que convierte los datos de la columna \code{.col_valores}.}
\usage{
get_tipo_cambio_df(
  .from_currency = "EUR",
  .to_currency = "USD",
  .from_date = as.Date(seq(as.Date(Sys.Date()), length = 2, by = "-1 month")[2]),
  .to_date = Sys.Date(),
  .frecuencia = "diaria",
  max_retries = 3,
  retry_delay = 2
)
}
\arguments{
\item{.from_currency}{Cadena de caracteres. C贸digo ISO a 3 letras de la divisa de base (base currency). Por defecto, \code{"EUR"}.}

\item{.to_currency}{Cadena de caracteres. C贸digo ISO a 3 letras de la divisa de cuenta (quote currency). Por defecto, \code{"USD"}.}

\item{.from_date}{Fecha. Fecha de inicio del dataframe a obtener. Por defecto, hoy hace un mes.}

\item{.to_date}{Fecha. Fecha de final del dataframe a obtener. Por defecto, hoy.}

\item{.frecuencia}{Frecuencia de los datos. Puede ser, \code{"diaria"}, \code{"mensual"} o \code{"anual"}. Por defecto, \code{"diaria"}.}
}
\description{
Funci贸n para obtener un dataframe con datos del tipo de cambio entre 2 divisas en un periodo especificado. Devuelve el dataframe inicial con una columna adicional \verb{tipo_cambio_".from_currency"_".to_currency"}, que convierte los datos de la columna \code{.col_valores}.
}
