% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_tipocambio.R
\name{get_tipo_cambio_plt}
\alias{get_tipo_cambio_plt}
\title{Función para obtener un gráfico con el tipo de cambio entre 2 divisas en una fecha específica.}
\usage{
get_tipo_cambio_plt(
  .from_currency = "EUR",
  .to_currency = "USD",
  .from_date = as.Date(seq(as.Date(Sys.Date()), length = 2, by = "-1 month")[2]),
  .to_date = Sys.Date(),
  .frecuencia = "diaria",
  .estatico = FALSE
)
}
\arguments{
\item{.from_currency}{Cadena de caracteres. Código ISO a 3 letras de la divisa de base (base currency). Por defecto, \code{"EUR"}.}

\item{.to_currency}{Cadena de caracteres. Código ISO a 3 letras de la divisa de cuenta (quote currency). Por defecto, \code{"USD"}.}

\item{.from_date}{Fecha. Fecha de inicio del gráfico a obtener. Por defecto, hoy hace un mes.}

\item{.to_date}{Fecha. Fecha de final del gráfico a obtener. Por defecto, hoy.}

\item{.frecuencia}{Frecuencia de los datos. Puede ser, \code{"diaria"}, \code{"mensual"} o \code{"anual"}. Por defecto, \code{"diaria"}.}

\item{.estatico}{Valor lógico. Establecer en \code{TRUE} si se desea que el gráfico no sea interactivo. Por defecto, \code{FALSE}.}
}
\description{
Función para obtener un gráfico con el tipo de cambio entre 2 divisas en una fecha específica.
}
