% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{graficos_estilo_victorgm}
\alias{graficos_estilo_victorgm}
\title{Fijar estilo de gráfico estándar para victorgm, con algunos parámetros opcionales}
\usage{
graficos_estilo_victorgm(
  .x,
  .tipo_grafico_x = "fecha",
  .tipo_grafico_y = NULL,
  .scale = 1,
  .accuracy = 1,
  .suffix = "",
  .minimo_eje_y = NULL,
  .maximo_eje_y = NULL,
  .date_labels = "\%Y",
  .angle_x_axis_labels = 45,
  .date_breaks = "1 year",
  .fecha_inicial_grafico = NULL,
  .fecha_final_grafico = NULL,
  .axis_position = "right",
  .legend_position = "bottom",
  .title_axis_size = 12,
  .text_axis_size = 12,
  .legend_text_size = 16,
  .text_size = 14,
  .nrows_legend_color = 3,
  .n_breaks_y = NULL,
  .paleta_utilizada = NULL,
  .n_colors = NULL,
  .title_x_axis = NULL,
  .title_y_axis = NULL,
  .title = NULL,
  .title_size = 18,
  .subtitle = NULL,
  .subtitle_size = 14,
  .caption = NULL,
  .caption_size = 10,
  .fuente_letra = "Segoe UI"
)
}
\arguments{
\item{.tipo_grafico_x}{Cadena de caracteres. Indica si se trata de una serie temporal o no. Puede tomar valores \code{"fecha"} o \code{"nofecha"}. Por defecto, \code{"fecha"}.}

\item{.tipo_grafico_y}{Cadena de caracteres. Sirve para indicar en ciertos casos la escala del eje y. Puede tomar valores \code{"millones"}, \code{"suffixing"}, \code{"porcentaje"} o \code{NULL}. Si es distinto a \code{NULL}, el resto de argumentos no se tienen en cuenta en la medida en que colisionen con los definidos para el tipo de gráfico. Por defecto, \code{NULL}.}

\item{.scale}{Número. Factor por el que dividir los números del eje. Por defecto, \code{1}. Para millones, p.ej., habría que fijar a 1e+6 o 1000000.}

\item{.accuracy}{Número. Número de decimales a mostrar en eje. Por defecto, \code{1}. Para mostrar 0 decimales, fijar a 1; para 1 decimal, fijar a 0.1; para 2, a 0.01 y así...}

\item{.suffix}{Cadena de caracteres. Por defecto, cadena vacía.}

\item{.minimo_eje_y}{Número. Valor mínimo en el eje. Por defecto, `NULL`` y el mínimo se adapta al gráfico}

\item{.maximo_eje_y}{Número. Valor máximo en el eje. Por defecto, \code{NULL} y el mínimo se adapta al gráfico}

\item{.date_labels}{Cadena de caracteres. Formato de las fechas en eje de abscisas. Por defecto, \code{"\%Y"}. Para mostrar el año y mes abreviado, fijar a \code{"\%Y-\%b"}; para fijar sólo el número del año \code{"\%Y"}. Para otros casos, ver documentación de \code{scale_x_date}.}

\item{.angle_x_axis_labels}{Número. Grado de inclinación de los datos en el eje horizontal. Sólo funciona con \code{.tipo_grafico_x = "fecha"}. Por defecto, \code{45}.}

\item{.date_breaks}{Cadena de caracteres. Frecuencia de las fechas en eje de abscisas. Por defecto, \code{"1 year"}. Para mostrar una fecha cada seis meses, fijar a \code{"6 months"}; para mostrar fechas cada 2 años, fijar a \code{"2 years"}.}

\item{.fecha_inicial_grafico}{Fecha. Sólo funciona con \code{.tipo_grafico_x = "fecha"}. Por defecto, \code{NULL}.}

\item{.fecha_final_grafico}{Fecha. Sólo funciona con \code{.tipo_grafico_x = "fecha"}. Por defecto, \code{NULL}.}

\item{.axis_position}{Cadena de caracteres. Posición del eje de ordenadas. Por defecto, \code{"right"}.}

\item{.legend_position}{Cadena de caracteres. Posición de la leyenda. Incluye la opción \code{"auto"}, que permite buscar un hueco "libre" dentro del gráfico y ponerlo en una esquina dentro del área de trazado. Por defecto, \code{"bottom"}.}

\item{.title_axis_size}{Número. Tamaño de la fuente en los títulos de los ejes. Sólo aplicable para \code{.tipo_grafico = "nofecha"}. Por defecto, \code{12}.}

\item{.text_axis_size}{Número. Tamaño de la fuente del texto de los ejes. Por defecto, \code{12}.}

\item{.legend_text_size}{Número. Tamaño de la fuente en la leyenda. Por defecto, \code{16}.}

\item{.text_size}{Número. Tamaño de la fuente en el resto del texto. Por defecto, \code{14}.}

\item{.nrows_legend_color}{Número. Número de filas a incluir en la leyenda. Por defecto, \code{3}.}

\item{.n_breaks_y}{Número. Número de breaks en el eje vertical. Por defecto, \code{NULL}.}

\item{.title_x_axis}{Título del eje x. Por defecto, \code{NULL}.}

\item{.title_y_axis}{Título del eje y. Por defecto, \code{NULL}.}

\item{.title}{Título del gráfico. Por defecto, \code{NULL}.}

\item{.subtitle}{Subtitulo del gráfico. Habitualmente usado para definir las unidades de medida. Por defecto, \code{NULL}.}

\item{.caption}{Nota al pie del gráfico. Por defecto, \code{NULL}.}

\item{.fuente_letra}{Cadena de caracteres. Fuente tipográfica a utilizar en el gráfico. Por defecto, utiliza la fuente: \code{"Segoe UI"}. Otras opciones comunes: \code{"Arial"}, \code{"Times"}, \code{"Courier"}, \code{"Helvetica"}, etc.}
}
\description{
Fijar estilo de gráfico estándar para victorgm, con algunos parámetros opcionales
}
