% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_strings.R
\name{limpiar_nombres}
\alias{limpiar_nombres}
\title{Te devuelve el input inicial sin tildes y sin espacios (y, en su caso, en minúsculas o mayúsculas). Basada en janitor::make_clean_names.}
\usage{
limpiar_nombres(
  .x,
  case = "snake",
  replace = c(`'` = "", `"` = "", `\%` = "_percent_", `#` = "_number_"),
  ascii = TRUE,
  use_make_names = TRUE,
  allow_dupes = FALSE,
  sep_in = "\\\\.",
  transliterations = "Latin-ASCII",
  parsing_option = 1,
  numerals = "asis",
  ...
)
}
\arguments{
\item{.x}{Input.}

\item{case}{The desired target case (default is \code{"snake"}) will be passed to \code{snakecase::to_any_case()} with the exception of "old_janitor", which exists only to support legacy code (it preserves the behavior of \code{clean_names()} prior to addition of the "case" argument (janitor versions <= 0.3.1). "old_janitor" is not intended for new code. See \code{to_any_case} for a wide variety of supported cases, including "sentence" and "title" case.}

\item{replace}{A named character vector where the name is replaced by the value.}

\item{ascii}{Convert the names to ASCII (\code{TRUE}, default) or not (\code{FALSE}).}

\item{use_make_names}{Should \code{make.names()} be applied to ensure that the output is usable as a name without quoting? (Avoiding \code{make.names()} ensures that the output is locale-independent but quoting may be required.)}

\item{allow_dupes}{Allow duplicates in the returned names (\code{TRUE}) or not (\code{FALSE}, the default).}

\item{sep_in}{(short for separator input) if character, is interpreted as a regular expression (wrapped internally into \code{stringr::regex()}). The default value is a regular expression that matches any sequence of non-alphanumeric values. All matches will be replaced by underscores (additionally to "_" and " ", for which this is always true, even if \code{NULL} is supplied). These underscores are used internally to split the strings into substrings and specify the word boundaries.}

\item{transliterations}{A character vector (if not \code{NULL}). The entries of this argument need to be elements of \code{stringi::stri_trans_list()} (like "Latin-ASCII", which is often useful) or names of lookup tables (currently only \code{"german"} is supported). In the order of the entries the letters of the input string will be transliterated via \code{stringi::stri_trans_general()} or replaced via the matches of the lookup table. When named character elements are supplied as part of 'transliterations', anything that matches the names is replaced by the corresponding value. You should use this feature with care in case of \code{case = "parsed"}, \code{case = "internal_parsing"} and \code{case = "none"}, since for upper case letters, which have transliterations/replacements of length 2, the second letter will be transliterated to lowercase, for example Oe, Ae, Ss, which might not always be what is intended. In this case you can make usage of the option to supply named elements and specify the transliterations yourself.}

\item{parsing_option}{An integer that will determine the parsing_option. "\n" 1: \code{"RRRStudio"} -> \code{"RRR_Studio"} "\n" 2: \code{"RRRStudio"} -> \code{"RRRS_tudio"} "\n" 3: \code{"RRRStudio"} -> \code{"RRRSStudio"}. This will become for example \code{"Rrrstudio"} when we convert to lower camel case. "\n"  -1, -2, -3: These \code{parsing_options}'s will suppress the conversion after non-alphanumeric values. "\n" 0: no parsing}

\item{numerals}{A character specifying the alignment of numerals (\code{"middle"}, \code{"left"}, \code{"right"}, \code{"asis"} or \code{"tight"}). I.e. \code{numerals = "left"} ensures that no output separator is in front of a digit.}
}
\description{
Te devuelve el input inicial sin tildes y sin espacios (y, en su caso, en minúsculas o mayúsculas). Basada en janitor::make_clean_names.
}
