% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_strings.R
\name{quitar_tildes}
\alias{quitar_tildes}
\title{Te devuelve el input inicial sin tildes (y, en su caso, en minúsculas o mayúsculas).}
\usage{
quitar_tildes(.x, tolower = FALSE, toupper = FALSE, quitar_espacios = FALSE)
}
\arguments{
\item{.x}{Input.}

\item{tolower}{Valor lógico. Si se establece en TRUE te convierte el input en minúsculas. Por defecto \code{FALSE}.}

\item{toupper}{Valor lógico. Si se establece en TRUE te convierte el input en mayúsculas. Por defecto \code{FALSE}.}

\item{quitar_espacios}{Valor lógico. Si se establece en TRUE te sustituye espacios y guiones por barras bajas. Por defecto \code{FALSE}.}
}
\description{
Te devuelve el input inicial sin tildes (y, en su caso, en minúsculas o mayúsculas).
}
