% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{tema_victorgm}
\alias{tema_victorgm}
\title{Tema ggplot2 de victorgm}
\usage{
tema_victorgm()
}
\value{
Un objeto theme de ggplot2.
}
\description{
Aplica el tema visual estándar de victorgm a un gráfico ggplot.
Este tema se caracteriza por un fondo transparente, líneas de cuadrícula horizontales punteadas, y la ausencia de líneas verticales y bordes de ejes innecesarios.
}
