% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{aplicar_paleta_victorgm}
\alias{aplicar_paleta_victorgm}
\title{Fijar estilo de gráfico estándar para VictorGM, con algunos parámetros opcionales}
\usage{
aplicar_paleta_victorgm(.x, .n_colors = NULL)
}
\description{
Fijar estilo de gráfico estándar para VictorGM, con algunos parámetros opcionales
}
