% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_fred.R
\name{get_fred_tbl}
\alias{get_fred_tbl}
\title{Obtiene una tabla con los datos de series temporales de FRED (St. Louis FED) adaptado en formato.}
\usage{
get_fred_tbl(
  .analysed_series_id,
  .analysed_observation_start = NULL,
  .analysed_observation_end = NULL,
  .analysed_frequency = NULL,
  .analysed_aggregation_method = NULL,
  .analysed_limit = NULL,
  .analysed_offset = NULL,
  .analysed_sort_order = NULL,
  .analysed_units = NULL,
  .analysed_realtime_start = NULL,
  .analysed_realtime_end = NULL,
  .analysed_vintage_dates = NULL,
  .analysed_output_type = NULL
)
}
\arguments{
\item{.analysed_series_id}{Requerido. Cadena de caracteres. Indica el código de la serie de la que se desea obtener la tabla. Se pueden consultar todas las series disponibles en la página web: https://fred.stlouisfed.org/.}

\item{.analysed_observation_start}{Cadena de caracteres. Primera observación de la serie temporal. Por defecto, NULL, devuelve desde la primera observación disponible.}

\item{.analysed_observation_end}{Fecha. Última observación de la serie temporal. Por defecto, NULL, devuelve desde la última observación disponible.}

\item{.analysed_frequency}{A}

\item{.analysed_aggregation_method}{a}

\item{.analysed_limit}{a}

\item{.analysed_offset}{a}

\item{.analysed_sort_order}{a}

\item{.analysed_units}{a}

\item{.analysed_realtime_start}{a}

\item{.analysed_realtime_end}{a}

\item{.analysed_vintage_dates}{a}

\item{.analysed_output_type}{a}
}
\description{
Obtiene una tabla con los datos de series temporales de FRED (St. Louis FED) adaptado en formato.
}
