% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_graficos.R
\name{mostrar}
\alias{mostrar}
\title{Muestra un objeto ggplot o gt, si existe en memoria, y con contenido dinámico si disponible. Además, si se está renderizando un pdf, hace una captura de la tabla gt con webshot2 y la muestra.}
\usage{
mostrar(.x, .estatico = FALSE, .width = 6, .height = 5)
}
\arguments{
\item{.x}{Objeto ggplot o gt a mostrar}

\item{.estatico}{Lógico. Si el objeto es un objeto ggplot, permite que se muestre como un gráfico interactivo en caso de establecerse en \code{FALSE}. Por defecto, \code{FALSE}.}

\item{.width}{Número. Ancho del gráfico a mostrar en pulgadas. Por defecto, \code{6}.}

\item{.height}{Número. Alto del gráfico a mostrar en pulgadas. Por defecto, \code{5}.}
}
\description{
Muestra un objeto ggplot o gt, si existe en memoria, y con contenido dinámico si disponible. Además, si se está renderizando un pdf, hace una captura de la tabla gt con webshot2 y la muestra.
}
